ChString bugfix of tolower and toupper function calls

+++ source/ChString/ChString.cpp
@@ -1,6 +1,7 @@
 #include "ChString.h"
 
 #include <sstream>
+#include <cctype>
 #include <regex>
 
 ChString::ChString() : data_("")
@@ -15,7 +16,7 @@ ChString::ChString(const ChString &other) : data_(other.data_)
 {
 }
 
-ChString::ChString(const char* str) : data_(str) 
+ChString::ChString(const char *str) : data_(str)
 {
 }
 
@@ -58,207 +59,226 @@ ChString::operator const std::string &() const
     return data_;
 }
 
-void ChString::fromNumber(int num) 
+void ChString::fromNumber(int num)
 {
     std::stringstream ss;
     ss << num;
     data_ = ss.str();
 }
 
-void ChString::fromNumber(float num) 
+void ChString::fromNumber(float num)
 {
     std::stringstream ss;
     ss << num;
     data_ = ss.str();
 }
 
-void ChString::fromNumber(double num) 
+void ChString::fromNumber(double num)
 {
     std::stringstream ss;
     ss << num;
     data_ = ss.str();
 }
 
-void ChString::fromNumber(unsigned int num) 
+void ChString::fromNumber(unsigned int num)
 {
     std::stringstream ss;
     ss << num;
     data_ = ss.str();
 }
 
-void ChString::fromNumber(long num) 
+void ChString::fromNumber(long num)
 {
     std::stringstream ss;
     ss << num;
     data_ = ss.str();
 }
 
-void ChString::fromNumber(unsigned long num) 
+void ChString::fromNumber(unsigned long num)
 {
     std::stringstream ss;
     ss << num;
     data_ = ss.str();
 }
 
-void ChString::fromNumber(long long num) 
+void ChString::fromNumber(long long num)
 {
     std::stringstream ss;
     ss << num;
     data_ = ss.str();
 }
 
-void ChString::fromNumber(unsigned long long num) 
+void ChString::fromNumber(unsigned long long num)
 {
     std::stringstream ss;
     ss << num;
     data_ = ss.str();
 }
 
-double ChString::toDouble(bool& valid) const 
+double ChString::toDouble(bool &valid) const
 {
     std::string str = data_;
 
     std::string::iterator end_pos = std::remove(str.begin(), str.end(), ' ');
     str.erase(end_pos, str.end());
 
-    try {
+    try
+    {
         size_t pos = 0;
         double result = std::stod(str, &pos);
         valid = (pos == str.length());
         return result;
-    } catch (...) {
+    }
+    catch (...)
+    {
         valid = false;
         return 0.0;
     }
 }
 
-int ChString::toInteger(bool& valid) const 
+int ChString::toInteger(bool &valid) const
 {
     std::string str = data_;
 
     std::string::iterator end_pos = std::remove(str.begin(), str.end(), ' ');
     str.erase(end_pos, str.end());
 
-    try {
+    try
+    {
         size_t pos = 0;
         int result = std::stoi(str, &pos);
         valid = (pos == str.length());
         return result;
-    } catch (...) {
+    }
+    catch (...)
+    {
         valid = false;
         return 0;
     }
 }
 
-float ChString::toFloat(bool& valid) const 
+float ChString::toFloat(bool &valid) const
 {
     std::string str = data_;
 
     std::string::iterator end_pos = std::remove(str.begin(), str.end(), ' ');
     str.erase(end_pos, str.end());
 
-    try {
+    try
+    {
         size_t pos = 0;
         float result = std::stof(str, &pos);
         valid = (pos == str.length());
         return result;
-    } catch (...) {
+    }
+    catch (...)
+    {
         valid = false;
         return 0.0f;
     }
 }
 
-long long ChString::toLongLong(bool& valid) const 
+long long ChString::toLongLong(bool &valid) const
 {
     std::string str = data_;
 
     std::string::iterator end_pos = std::remove(str.begin(), str.end(), ' ');
     str.erase(end_pos, str.end());
 
-    try {
+    try
+    {
         size_t pos = 0;
         long long result = std::stoll(str, &pos);
         valid = (pos == str.length());
         return result;
-    } catch (...) {
+    }
+    catch (...)
+    {
         valid = false;
         return 0;
     }
 }
 
-unsigned int ChString::toUnsignedInteger(bool& valid) const 
+unsigned int ChString::toUnsignedInteger(bool &valid) const
 {
     std::string str = data_;
 
     std::string::iterator end_pos = std::remove(str.begin(), str.end(), ' ');
     str.erase(end_pos, str.end());
 
-    try {
+    try
+    {
         size_t pos;
         unsigned int result = std::stoul(str, &pos);
         valid = pos == str.length();
         return result;
     }
-    catch (...) {
+    catch (...)
+    {
         valid = false;
         return 0;
     }
 }
 
-unsigned long ChString::toUnsignedLong(bool& valid) const 
+unsigned long ChString::toUnsignedLong(bool &valid) const
 {
     std::string str = data_;
 
     std::string::iterator end_pos = std::remove(str.begin(), str.end(), ' ');
     str.erase(end_pos, str.end());
 
-    try {
+    try
+    {
         size_t pos;
         unsigned long result = std::stoul(str, &pos);
         valid = pos == str.length();
         return result;
     }
-    catch (...) {
+    catch (...)
+    {
         valid = false;
         return 0;
     }
 }
 
-unsigned long long ChString::toUnsignedLongLong(bool& valid) const 
+unsigned long long ChString::toUnsignedLongLong(bool &valid) const
 {
     std::string str = data_;
 
     std::string::iterator end_pos = std::remove(str.begin(), str.end(), ' ');
     str.erase(end_pos, str.end());
 
-    try {
+    try
+    {
         size_t pos;
         unsigned long long result = std::stoull(str, &pos);
         valid = pos == str.length();
         return result;
     }
-    catch (...) {
+    catch (...)
+    {
         valid = false;
         return 0;
     }
 }
 
-ChString ChString::concat(const ChString& str) const 
+ChString ChString::concat(const ChString &str) const
 {
     return ChString(data_ + str.getData());
 }
 
-ChString ChString::concat(const std::string& str) const 
+ChString ChString::concat(const std::string &str) const
 {
     return ChString(data_ + str);
 }
 
-ChString ChString::substring(int start, int size) const 
+ChString ChString::substring(int start, int size) const
 {
     size_t size_ = static_cast<size_t>(size);
     size_t start_ = static_cast<size_t>(start);
 
-    if (size_ == 0) {
+    if (size_ == 0)
+    {
         size_ = data_.length() - start_;
     }
 
@@ -296,10 +316,11 @@ ChString ChString::removeAll(char delim)
     return ChString(value);
 }
 
-ChString ChString::removeFirst(ChString str) const 
+ChString ChString::removeFirst(ChString str) const
 {
     size_t pos = data_.find(str.data_);
-    if (pos == std::string::npos) {
+    if (pos == std::string::npos)
+    {
         return *this;
     }
     ChString result(data_);
@@ -307,16 +328,17 @@ ChString ChString::removeFirst(ChString str) const
     return result;
 }
 
-ChString ChString::removeFirst(std::string str) const 
+ChString ChString::removeFirst(std::string str) const
 {
     ChString chStr(str);
     return removeFirst(chStr);
 }
 
-ChString ChString::removeLast(ChString str) const 
+ChString ChString::removeLast(ChString str) const
 {
     size_t pos = data_.rfind(str.data_);
-    if (pos == std::string::npos) {
+    if (pos == std::string::npos)
+    {
         return *this;
     }
     ChString result(data_);
@@ -324,22 +346,26 @@ ChString ChString::removeLast(ChString str) const
     return result;
 }
 
-ChString ChString::removeLast(std::string str) const 
+ChString ChString::removeLast(std::string str) const
 {
     ChString chStr(str);
     return removeLast(chStr);
 }
 
-char ChString::first() const {
+char ChString::first() const
+{
     return data_.empty() ? '\0' : data_.front();
 }
 
-char ChString::last() const {
+char ChString::last() const
+{
     return data_.empty() ? '\0' : data_.back();
 }
 
-char ChString::popFirst() {
-    if (data_.empty()) {
+char ChString::popFirst()
+{
+    if (data_.empty())
+    {
         return '\0';
     }
     char c = data_.front();
@@ -347,8 +373,10 @@ char ChString::popFirst() {
     return c;
 }
 
-char ChString::popLast() {
-    if (data_.empty()) {
+char ChString::popLast()
+{
+    if (data_.empty())
+    {
         return '\0';
     }
     char c = data_.back();
@@ -356,191 +384,242 @@ char ChString::popLast() {
     return c;
 }
 
-bool ChString::contains(const ChString& value, bool isCaseSensitive) const {
+bool ChString::contains(const ChString &value, bool isCaseSensitive) const
+{
     std::string str = data_;
     std::string subStr = value.getData();
-    if (!isCaseSensitive) {
+    if (!isCaseSensitive)
+    {
         std::transform(str.begin(), str.end(), str.begin(), ::tolower);
         std::transform(subStr.begin(), subStr.end(), subStr.begin(), ::tolower);
     }
     return str.find(subStr) != std::string::npos;
 }
 
-bool ChString::contains(const std::string& value, bool isCaseSensitive) const {
+bool ChString::contains(const std::string &value, bool isCaseSensitive) const
+{
     std::string str = data_;
     std::string subStr = value;
-    if (!isCaseSensitive) {
+    if (!isCaseSensitive)
+    {
         std::transform(str.begin(), str.end(), str.begin(), ::tolower);
         std::transform(subStr.begin(), subStr.end(), subStr.begin(), ::tolower);
     }
     return str.find(subStr) != std::string::npos;
 }
 
-int ChString::count(const ChString& value, bool isCaseSensitive) const {
+int ChString::count(const ChString &value, bool isCaseSensitive) const
+{
     std::string str = data_;
     std::string subStr = value.getData();
     int count = 0;
     size_t pos = 0;
-    if (!isCaseSensitive) {
+    if (!isCaseSensitive)
+    {
         std::transform(str.begin(), str.end(), str.begin(), ::tolower);
         std::transform(subStr.begin(), subStr.end(), subStr.begin(), ::tolower);
     }
-    while ((pos = str.find(subStr, pos)) != std::string::npos) {
+    while ((pos = str.find(subStr, pos)) != std::string::npos)
+    {
         ++count;
         pos += subStr.length();
     }
     return count;
 }
 
-int ChString::count(const std::string& value, bool isCaseSensitive) const {
+int ChString::count(const std::string &value, bool isCaseSensitive) const
+{
     std::string str = data_;
     std::string subStr = value;
     int count = 0;
     size_t pos = 0;
-    if (!isCaseSensitive) {
+    if (!isCaseSensitive)
+    {
         std::transform(str.begin(), str.end(), str.begin(), ::tolower);
         std::transform(subStr.begin(), subStr.end(), subStr.begin(), ::tolower);
     }
-    while ((pos = str.find(subStr, pos)) != std::string::npos) {
+    while ((pos = str.find(subStr, pos)) != std::string::npos)
+    {
         ++count;
         pos += subStr.length();
     }
     return count;
 }
 
-ChString ChString::toLower() const {
+ChString ChString::toLower() const
+{
     std::string result = data_;
-    std::transform(result.begin(), result.end(), result.begin(), [](unsigned char c) {
-        return std::tolower(c);
-    });
+    std::transform(result.begin(), result.end(), result.begin(), [](unsigned char c)
+                   { return std::tolower(c); });
     return result;
 }
 
-ChString ChString::toUpper() const {
+ChString ChString::toUpper() const
+{
     std::string result = data_;
-    std::transform(result.begin(), result.end(), result.begin(), [](unsigned char c) {
-        return std::toupper(c);
-    });
+    std::transform(result.begin(), result.end(), result.begin(), [](unsigned char c)
+                   { return std::toupper(c); });
     return result;
 }
 
-ChString ChString::capitalize() const {
+ChString ChString::capitalize() const
+{
     std::string result = data_;
     bool newWord = true;
-    for (auto& c : result) {
-        if (std::isalpha(c)) {
-            if (newWord) {
+    for (auto &c : result)
+    {
+        if (std::isalpha(c))
+        {
+            if (newWord)
+            {
                 c = std::toupper(c);
                 newWord = false;
             }
-        } else {
+        }
+        else
+        {
             newWord = true;
         }
     }
     return result;
 }
 
-bool ChString::beginsWith(const ChString& str) const {
+bool ChString::beginsWith(const ChString &str) const
+{
     return data_.find(str.data_) == 0;
 }
 
-bool ChString::beginsWith(const std::string& str) const {
+bool ChString::beginsWith(const std::string &str) const
+{
     return data_.find(str) == 0;
 }
 
-bool ChString::beginsWith(char ch) const {
-    if (data_.empty()) {
+bool ChString::beginsWith(char ch) const
+{
+    if (data_.empty())
+    {
         return false;
     }
     return data_.at(0) == ch;
 }
 
-bool ChString::endsWith(const ChString& str) const {
-    if (data_.length() < str.data_.length()) {
+bool ChString::endsWith(const ChString &str) const
+{
+    if (data_.length() < str.data_.length())
+    {
         return false;
     }
     return data_.compare(data_.length() - str.data_.length(), str.data_.length(), str.data_) == 0;
 }
 
-bool ChString::endsWith(const std::string& str) const {
-    if (data_.length() < str.length()) {
+bool ChString::endsWith(const std::string &str) const
+{
+    if (data_.length() < str.length())
+    {
         return false;
     }
     return data_.compare(data_.length() - str.length(), str.length(), str) == 0;
 }
 
-bool ChString::endsWith(char ch) const {
-    if (data_.empty()) {
+bool ChString::endsWith(char ch) const
+{
+    if (data_.empty())
+    {
         return false;
     }
     return data_.at(data_.length() - 1) == ch;
 }
 
-ChString ChString::erase(int first, int last) const {
-    if (first >= data_.length() || first > last) {
+ChString ChString::erase(int first, int last) const
+{
+    if (static_cast<size_t>(first) >= data_.length() || first > last)
+    {
         return ChString("");
     }
     return ChString(data_.substr(0, first) + data_.substr(last + 1));
 }
 
-ChString ChString::leftJustified(size_t width, char fill, bool truncate) const {
-    if (data_.size() >= width) {
-        if (truncate) {
+ChString ChString::leftJustified(size_t width, char fill, bool truncate) const
+{
+    if (data_.size() >= width)
+    {
+        if (truncate)
+        {
             return ChString(data_.substr(0, width));
-        } else {
+        }
+        else
+        {
             return ChString(data_);
         }
-    } else {
+    }
+    else
+    {
         return ChString(data_ + std::string(width - data_.size(), fill));
     }
 }
 
-ChString ChString::rightJustified(size_t width, char fill, bool truncate) const {
-    if (data_.size() >= width) {
-        if (truncate) {
+ChString ChString::rightJustified(size_t width, char fill, bool truncate) const
+{
+    if (data_.size() >= width)
+    {
+        if (truncate)
+        {
             return ChString(data_.substr(data_.size() - width, width));
-        } else {
+        }
+        else
+        {
             return ChString(data_);
         }
-    } else {
+    }
+    else
+    {
         return ChString(std::string(width - data_.size(), fill) + data_);
     }
 }
 
-std::list<ChString> ChString::split(const ChString& separator, bool keepEmptyParts, bool caseSensitive) const {
+std::list<ChString> ChString::split(const ChString &separator, bool keepEmptyParts, bool caseSensitive) const
+{
     std::list<ChString> result;
     size_t pos = 0;
     size_t separatorLength = separator.data_.length();
     std::string data = data_;
     std::string sep = separator.data_;
-    while (pos <= data.length() - separatorLength) {
-        if ((caseSensitive && data.substr(pos, separatorLength) == sep) || (!caseSensitive && ChString(data.substr(pos, separatorLength)).toLower().data_ == separator.toLower().data_)) {
+    while (pos <= data.length() - separatorLength)
+    {
+        if ((caseSensitive && data.substr(pos, separatorLength) == sep) || (!caseSensitive && ChString(data.substr(pos, separatorLength)).toLower().data_ == separator.toLower().data_))
+        {
             result.push_back(data.substr(0, pos));
             data = data.erase(0, pos + separatorLength - 1);
             pos = 0;
         }
-        else {
+        else
+        {
             ++pos;
         }
     }
-    if (keepEmptyParts || !data.empty()) {
+    if (keepEmptyParts || !data.empty())
+    {
         result.push_back(*this);
     }
     return result;
 }
 
-std::list<ChString> ChString::split(const std::string& separator, bool keepEmptyParts, bool caseSensitive) const {
+std::list<ChString> ChString::split(const std::string &separator, bool keepEmptyParts, bool caseSensitive) const
+{
     return split(ChString(separator), keepEmptyParts, caseSensitive);
 }
 
-std::list<ChString> ChString::split(char separator, bool keepEmptyParts, bool caseSensitive) const {
+std::list<ChString> ChString::split(char separator, bool keepEmptyParts, bool caseSensitive) const
+{
     return split(ChString(1, separator), keepEmptyParts, caseSensitive);
 }
 
-size_t ChString::size() const {
+size_t ChString::size() const
+{
     return data_.size();
 }
 
-const std::string& ChString::getData() const {
+const std::string &ChString::getData() const
+{
     return data_;
 }
